+++ src/cli/steadybit.ts
@@ -12,6 +12,7 @@ new Command()
   // eslint-disable-next-line
   .version(require('../../package.json').version)
   .command('config', 'Show/modify the CLI configuration and authentication profiles.')
+  .command('def-repo', 'Change tags and verify a task/policy definition repository state.')
   .command('service', 'Alias for the "service-definition" command.')
   .command('service-definition', 'Configure or verify service definitions.')
   .parseAsync(process.argv);